requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
diff-match-patch==20121119
django-audit-log==0.7.0
django-braces==1.8.0
django-cors-headers==0.13
django-debug-toolbar==1.2.2
django-import-export==0.2.8
django-mockups==0.4.8
django-oauth-toolkit==0.8.1
djangorestframework==2.4.4
easy-thumbnails==2.2
ipdb==0.8
ipython==2.3.1
oauthlib==0.7.2
psycopg2==2.5.4
shortuuid==0.4.2
six==1.9.0
sorl-thumbnail==11.12.1b
sqlparse==0.1.14
tablib==0.10.0
wsgiref==0.1.2
xlwt==1.0.0
